pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "envolees"
version = "0.1.0"
description = "Backtest engine for Donchian breakout strategy with prop firm simulation"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Raph" }
]
keywords = ["trading", "backtest", "prop-firm", "donchian", "ftmo", "gft"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "yfinance>=0.2.40",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pandas-stubs>=2.2.0",
]

[project.scripts]
envolees = "envolees.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["envolees"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["numpy.typing.mypy_plugin"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=envolees --cov-report=term-missing"
